+++ hazimp_preprocessing/curve_data/create_vuln_xml.py
@@ -151,8 +151,11 @@ def csv_curve2nrml(csv_filename, xml_filename):
             for iml in imls:
                 if numpy.isnan(iml):
                     continue
-                loss_ratio += str(row[str(int(iml))]) + ' '
-                coef_var += str((float(row['Alpha'])/float(row['Beta']))) + ' '
+                loss_ratio += str(row[str(iml)]) + ' '
+                if (row['Alpha'] == '') | (row['Beta'] == ''):
+                    coef_var += '0.0 '
+                else:
+                    coef_var += str((float(row['Alpha'])/float(row['Beta']))) + ' '
             write_nrml_curve(xml_h, row['vulnerabilityFunctionID'],
                              imls, csv_dict['IMT'][0],
                              loss_ratio, coef_var)